requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
ujson==5.3.0
pymongo==4.1.1
numba==0.55.2
requests==2.27.1
python-dotenv==0.20.0